BAVL-612: Refactoring it all to use shared functions and common patterns
